refactor: cambiar tipo de dato de documento a long en las entidades Acudientes, Estudiantes, Personas y Profesores

+++ demo/src/main/java/com/example/ControAcceso/Model/Acudientes.java
@@ -13,7 +13,7 @@ public class Acudientes extends Personas {
 
     }
 
-    public Acudientes(int documento, String tipoDocumento,
+    public Acudientes(long documento, String tipoDocumento,
             String primerNombre, String segundoNombre,
             String primerApellido, String segundoApellido,
             String direccion, long telefono,
